큐리 스킬 버그 수정1
1. 데미지가 104로 들어가는 문제는 데미지가 104.9999...인 상황에서 int 캐스팅을 했기 때문. 반올림을 하도록 수정.
2. '가연성 부착물'이 주 대상에게 2회 적용되는 문제 수정

+++ Assets/Scripts/Battle/Damage/DamageCalculator.cs
@@ -277,7 +277,7 @@ public class DamageCalculator
 	public static float GetActualDamage(SkillInstanceData skillInstanceData, bool isHealth)
 	{
 		float actualDamage = skillInstanceData.GetDamage().resultDamage;
-		int finalDamage = 0; // 최종 대미지 (정수로 표시되는)
+		float finalDamage = 0; // 최종 대미지 (정수로 표시되는)
         Skill appliedSkill = skillInstanceData.GetSkill();
         Unit target = skillInstanceData.GetMainTarget();
         Unit caster = skillInstanceData.GetCaster();
@@ -329,7 +329,7 @@ public class DamageCalculator
 				// actualDamage = actualDamage;
 			}
 
-			finalDamage = (int) actualDamage;
+			finalDamage = actualDamage;
 
 			// 보호막에 따른 대미지 삭감 - 실제 실드는 깎이지 않음
 			if (target.HasStatusEffect(StatusEffectType.Shield))
@@ -354,7 +354,7 @@ public class DamageCalculator
 			}
 		}
 		
-		return (float) finalDamage;
+		return finalDamage;
 	}	
 }
 }
